move ProposalSearch under the Query namespace
